requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.3.0
Werkzeug==2.2.2
